changed list parsing and moved set next hackathon
