+++ hms-api/dev/http/http-client.env.json
@@ -1,7 +1,7 @@
 {
   "local": {
     "host": "http://localhost:4566",
-    "api_id": "08fptbt9hv",
+    "api_id": "2go1i21sac",
     "hackathon_id": "4eb2d486-c786-431e-a4fd-4c093ed30642",
     "category_id": "7f38993d-073a-4932-af79-07440012f268",
     "category_id_2": "a8ce5468-ec97-4b99-acea-511c54817956",
