+++ hms-app/.env
@@ -1,1 +1,1 @@
-REACT_APP_CORE_URL=http://localhost:4566/restapis/08fptbt9hv/local/_user_request_
+REACT_APP_CORE_URL=http://localhost:4566/restapis/2go1i21sac/local/_user_request_
